code review comment tweaks
